# api/openapi/ops-platform.yaml
openapi: 3.0.3
info:
  title: Ops Platform API
  version: 0.1.0
servers:
  - url: /api/v1
paths:
...
  /namespaces/{namespace}/pods/{pod}:
    get:
      summary: Get pod diagnostic detail
      description: Returns runtime state, current and previous container termination details, owner references, resolved controller chain, conditions and scheduling/network metadata.
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/Namespace"
        - $ref: "#/components/parameters/Pod"
      responses:
        "200":
          description: Pod detail
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/ResponseEnvelope"
                  - type: object
                    properties:
                      data:
                        $ref: "#/components/schemas/PodDetail"
    delete:
      summary: Delete pod
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/Namespace"
        - $ref: "#/components/parameters/Pod"
        - name: confirm
          in: query
          required: true
          schema:
            type: boolean
      responses:
        "200":
          description: Delete request submitted
...
  /namespaces/{namespace}/pods/{pod}/restart:
    post:
      summary: Restart a controller-managed pod by deleting it for recreation
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/Namespace"
        - $ref: "#/components/parameters/Pod"
        - name: confirm
          in: query
          required: true
          schema:
            type: boolean
      responses:
        "200":
          description: Managed pod restart submitted
...
  /namespaces/{namespace}/deployments/{name}/scale:
    post:
      summary: Scale deployment
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/Namespace"
        - $ref: "#/components/parameters/Name"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - replicas
              properties:
                replicas:
                  type: integer
                  minimum: 0
                  maximum: 100
      responses:
        "200":
          description: Deployment scaled
  /namespaces/{namespace}/deployments/{name}/restart:
    post:
      summary: Restart deployment by annotating pod template
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/Namespace"
        - $ref: "#/components/parameters/Name"
      responses:
        "200":
          description: Deployment restart submitted
...
  /namespaces/{namespace}/statefulsets/{name}/scale:
    post:
      summary: Scale statefulset
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/Namespace"
        - $ref: "#/components/parameters/Name"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - replicas
              properties:
                replicas:
                  type: integer
                  minimum: 0
                  maximum: 100
      responses:
        "200":
          description: StatefulSet scaled
  /namespaces/{namespace}/statefulsets/{name}/restart:
    post:
      summary: Roll out a statefulset restart by annotating its pod template
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/Namespace"
        - $ref: "#/components/parameters/Name"
      responses:
        "200":
          description: StatefulSet restart submitted
...
  /namespaces/{namespace}/daemonsets/{name}/restart:
    post:
      summary: Roll out a daemonset restart by annotating its pod template
      description: DaemonSets using the OnDelete update strategy return HTTP 409 and must have their managed Pods restarted individually.
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/Namespace"
        - $ref: "#/components/parameters/Name"
      responses:
        "200":
          description: DaemonSet restart submitted
        "409":
          description: DaemonSet uses the OnDelete update strategy
...
  /namespaces/{namespace}/cronjobs/{name}/suspend:
    post:
      summary: Suspend future cronjob schedules
      description: Sets spec.suspend=true. Already-started Jobs are not stopped. Requires operator/admin, explicit confirmation, and audit.
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/Namespace"
        - $ref: "#/components/parameters/Name"
        - name: confirm
          in: query
          required: true
          schema:
            type: boolean
      responses:
        "200":
          description: CronJob suspended
  /namespaces/{namespace}/cronjobs/{name}/resume:
    post:
      summary: Resume future cronjob schedules
      description: Sets spec.suspend=false. Missed schedules are handled by Kubernetes according to startingDeadlineSeconds. Requires operator/admin, explicit confirmation, and audit.
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/Namespace"
        - $ref: "#/components/parameters/Name"
        - name: confirm
          in: query
          required: true
          schema:
            type: boolean
      responses:
        "200":
          description: CronJob resumed
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    Namespace:
      name: namespace
      in: path
      required: true
      schema:
        type: string
    Pod:
      name: pod
      in: path
      required: true
      schema:
        type: string
    Name:
      name: name
      in: path
      required: true
      schema:
        type: string
...
  schemas:
    ResponseEnvelope:
      type: object
      required:
        - code
        - message
        - data
      properties:
        code:
          type: integer
          format: int32
        message:
          type: string
        data:
          nullable: true
...
    OwnerReference:
      type: object
      required:
        - kind
        - name
        - controller
      properties:
        kind:
          type: string
        name:
          type: string
        controller:
          type: boolean
    ContainerSummary:
      type: object
      properties:
        name:
          type: string
        image:
          type: string
        ready:
          type: boolean
        restart_count:
          type: integer
          format: int32
        state:
          type: string
        reason:
          type: string
        exit_code:
          type: integer
          format: int32
        started_at:
          type: string
          format: date-time
        finished_at:
          type: string
          format: date-time
        last_state:
          type: string
        last_reason:
          type: string
        last_exit_code:
          type: integer
          format: int32
        last_finished_at:
          type: string
          format: date-time
    PodSummary:
      type: object
      properties:
        namespace:
          type: string
        name:
          type: string
        phase:
          type: string
        status:
          type: string
        node_name:
          type: string
        ready:
          type: boolean
        restart_count:
          type: integer
          format: int32
        containers:
          type: array
          items:
            $ref: "#/components/schemas/ContainerSummary"
        created_at:
          type: string
          format: date-time
    PodCondition:
      type: object
      properties:
        type:
          type: string
        status:
          type: string
        reason:
          type: string
        message:
          type: string
    PodDetail:
      allOf:
        - $ref: "#/components/schemas/PodSummary"
        - type: object
          properties:
            init_containers:
              type: array
              items:
                $ref: "#/components/schemas/ContainerSummary"
            pod_ip:
              type: string
            host_ip:
              type: string
            qos_class:
              type: string
            service_account:
              type: string
            restart_policy:
              type: string
            start_time:
              type: string
              format: date-time
            labels:
              type: object
              additionalProperties:
                type: string
            annotations:
              type: object
              additionalProperties:
                type: string
            owner_refs:
              type: array
              items:
                $ref: "#/components/schemas/OwnerReference"
            controller_chain:
              type: array
              items:
                $ref: "#/components/schemas/OwnerReference"
            conditions:
              type: array
              items:
                $ref: "#/components/schemas/PodCondition"
...